test: fix something
