+++ .mocharc.yml
@@ -1,0 +1,3 @@
+exit: true
+file:
+  - ./src/test/bootstrap.js
