+++ .travis.yml
@@ -14,11 +14,9 @@ branches:
 
 install:
   - npm ci
+  - cp .env.example .env
 
-script:
-  - npm run check-package
-  - npm run build:prod
-  - npm run test
+script: npm run check-package && npm run lint && npm run test
 
 before_deploy:
   # Add token for @dashevo private npm registry
