+++ karma.conf.js
@@ -1,17 +1,34 @@
-const webpackConfig = require("./webpack.base.config");
+/* eslint-disable import/no-extraneous-dependencies */
+const webpack = require('webpack');
+const dotenvResult = require('dotenv-safe').config();
+
+const webpackBaseConfig = require("./webpack.base.config");
+
+
+if (dotenvResult.error) {
+  throw dotenvResult.error;
+}
 
 module.exports = (config) => {
   config.set({
     frameworks: ['mocha', 'chai'],
     files: [
-      'src/index.ts',
+      'src/**/*.spec.ts',
       'tests/functional/sdk.js',
     ],
     preprocessors: {
-      'src/index.ts': ['webpack'],
+      'src/**/*.spec.ts': ['webpack'],
       'tests/functional/sdk.js': ['webpack'],
     },
-    webpack: {...webpackConfig, target:'web'},
+    webpack: {
+      ...webpackBaseConfig,
+      mode: 'development',
+      plugins: [
+        new webpack.EnvironmentPlugin(
+          dotenvResult.parsed,
+        ),
+      ],
+    },
     reporters: ['mocha'],
     port: 9876,
     colors: true,
