+++ package.json
@@ -4,19 +4,17 @@
   "description": "Dash library for JavaScript/TypeScript ecosystem (Wallet, DAPI, Primitives, BLS, ...)",
   "main": "dist/dash.cjs.js",
   "unpkg": "dist/dash.min.js",
-  "browser": "dist/dash.min.js",
   "types": "dist/src/index.d.ts",
   "scripts": {
-    "start:dev": "nodemon --exec 'npm run build:dev && npm run test:ts'",
-    "build:dev": "webpack -d --display-error-details --progress",
-    "build:prod": "webpack -p --display-error-details --progress",
+    "start:dev": "nodemon --exec 'npm run build && npm run test:unit'",
+    "build": "webpack --display-error-details",
     "check-package": "npm run check-package:name && npm run check-package:version",
     "check-package:name": "test $(jq -r .name package.json) = $(jq -r .name package-lock.json)",
     "check-package:version": "test $(jq -r .version package.json) = $(jq -r .version package-lock.json)",
-    "test": "npm run test:ts && npm run build:prod && npm run test:js && npm run test:browsers",
+    "test": "npm run test:unit && npm run build && npm run test:functional && npm run test:browsers",
     "test:browsers": "karma start ./karma.conf.js --single-run",
-    "test:ts": "mocha -r ts-node/register \"src/**/*.spec.ts\"",
-    "test:js": "mocha --recursive tests/**/*.js"
+    "test:unit": "mocha -r ts-node/register \"src/**/*.spec.ts\"",
+    "test:functional": "mocha --recursive tests/functional/**/*.js"
   },
   "repository": {
     "type": "git",
@@ -56,6 +54,7 @@
     "ts-node": "^8.10.2",
     "typescript": "^3.9.5",
     "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11"
+    "webpack-cli": "^3.3.11",
+    "webpack-node-externals": "^1.7.2"
   }
 }
