+++ src/test/bootstrap.js
@@ -1,0 +1,7 @@
+const dotenvSafe = require('dotenv-safe');
+
+const path = require('path');
+
+dotenvSafe.config({
+  path: path.resolve(__dirname, '..', '..', '.env'),
+});
